data flow with hooks
