+++ package.json
@@ -14,6 +14,7 @@
     "clsx": "^1.1.1",
     "d3-hierarchy": "^1.1.9",
     "express": "^4.17.1",
+    "lodash.debounce": "^4.0.8",
     "moment": "^2.26.0",
     "moment-timezone": "^0.5.31",
     "node-sass": "^4.14.1",
@@ -30,6 +31,7 @@
     "react-scripts": "3.4.1",
     "react-share": "^4.2.0",
     "react-zoom-pan-pinch": "^1.6.1",
+    "use-async-memo": "^1.2.2",
     "wikidata-sdk": "^7.6.0"
   },
   "scripts": {
