+++ src/App.scss
@@ -23,6 +23,9 @@
   .dropdown-toggle::after {
     vertical-align: middle;
   }
+  .navbar-brand {
+    margin-right: 0;
+  }
 
   &.light {
   }
