+++ src/components/Node/Node.js
@@ -40,7 +40,6 @@ export default function Node({
   toggleSpouses,
   setFocusedNode,
   focusedNode,
-  reloadTreeFromFocused,
   debug,
 }) {
   if (debug) console.log(node);
@@ -76,6 +75,8 @@ export default function Node({
         [gender]: gender,
       })}
       onClick={() => setFocusedNode(node)}
+      //data-id={node.data.id}
+      //data-tree-id={node.treeId}
     >
       <div
         className="imgWrapper"
@@ -305,13 +306,7 @@ export default function Node({
             </span>
           </Button>
         )}
-      {showModal && (
-        <DetailsModal
-          hideModal={hideModal}
-          node={node}
-          reloadTreeFromFocused={reloadTreeFromFocused}
-        />
-      )}
+      {showModal && <DetailsModal hideModal={hideModal} node={node} />}
     </div>
   );
 }
